+++ tests/test_threading.py
@@ -9,7 +9,7 @@ from scru128 import scru128, Scru128Id
 
 class TestThreading(unittest.TestCase):
     def test_threading(self) -> None:
-        """Generates no IDs sharing same timestamp and counter under multithreading"""
+        """Generates no IDs sharing same timestamp and counters under multithreading"""
 
         def producer(q: queue.Queue[Scru128Id]) -> None:
             for i in range(10000):
@@ -22,6 +22,6 @@ class TestThreading(unittest.TestCase):
         s = set()
         while not (q.empty() and threading.active_count() < 2):
             e = q.get()
-            s.add((e.timestamp, e.counter))
+            s.add((e.timestamp, e.counter_hi, e.counter_lo))
 
         self.assertEqual(len(s), 4 * 10000)